Opea/requirements.txt
pyspark==3.5.1
delta-spark==3.2.0
boto3==1.34.46
botocore==1.34.46
pandas==2.1.4
openpyxl==3.1.2
pyarrow==14.0.2
python-dotenv==1.0.1
pytest==7.4.4
